Revert "stepper_buzz: Move STEPPER_BUZZ command from stepper.py to new file"
This reverts commit 7f1591d0c589a0e8a45805fb3b4f3152d5fa4362.
